app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
sqlalchemy_utils
python-jose[cryptography]
passlib[bcrypt]
python-multipart
